pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comboi"
version = "0.1.0"
description = "Serverless DuckDB-based medallion ETL system"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "duckdb>=0.10.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "adlfs>=2024.4.0",
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.8.0",
    "azure-storage-queue>=12.10.0",
    "azure-functions>=1.15.0",
    "structlog>=24.1.0",
    "splink>=3.10.3",
    "pandas>=2.2.0",
    "tenacity>=8.2.3",
    "typer>=0.12.3",
    "dbt-core>=1.7.0",
    "dbt-duckdb>=1.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
    "pre-commit>=3.3.3",
]

[project.scripts]
comboi = "comboi.cli:app"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=comboi --cov-report=html --cov-report=term-missing"

